tweeking my third implementation for MSVC9 which will hopefully remove the warnings I getting
SVN-Revision: 63990

+++ SOC/2010/bit_masks/boost/integer/high_bits_mask.hpp
@@ -30,7 +30,7 @@ namespace detail {
         BOOST_STATIC_CONSTANT(unsigned_type, val = (~( unsigned_type(0)))
               &
             (~(mpl::shift_right<
-                mpl::size_t<~( T(0) )>,
+                mpl::size_t<static_cast<unsigned int>(~(T(0)))>,
                 mpl::size_t<Width>
             >::type::value))
         );
